backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
alembic==1.12.0
yfinance==0.2.18
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
redis==4.6.0
celery==5.3.1
numpy==1.24.3
pandas==2.0.3